+++ gems/aws-sdk-core/lib/aws-sdk-core/binary.rb
@@ -1,0 +1,3 @@
+require_relative 'binary/decode_handler'
+require_relative 'binary/event_stream_decoder'
+require_relative 'binary/event_parser'